Nice screen done

+++ main.sh
@@ -1,4 +1,13 @@
 #!/usr/bin/bash
+#checks if kdialog is installed or not
+installFlag=0 
+if ! command -v kdialog &> /dev/null
+then
+   echo "a key command was not found on your system ,Please Install kdialog and rerun the script"
+   exit
+fi
+
+
 clear
 kdialog --menu "Operations Menu" "1" "Create Table" "2" "List Tables" "3" "Drop Table" "4" "Insert into Table" "5" "Select From Table" "6" "Delete From Table" "7" "Update Table" "0" "Exit" > out
